Moved new scatter from hail.experimental to hail.plot. * Renamed fields to be as close as possible to previous scatter plot names * Added marker size option
Removed unused import +  space from experimental/plots
